tests/wpt/tests/performance-timeline/tentative/performance-entry-source-deleted-frame.html

<!DOCTYPE html>

<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>

<body>
</body>
<script>
  promise_test(async () => {
    // Create child iframe.
    const childFrame = document.createElement('iframe');
    childFrame.src = '../resources/child-frame.html';

    let childEntry;
    // Load child frame.
    const loadChildFramePromise = new Promise(resolve => {
      childFrame.addEventListener('load', () => {
        childEntry = performance.getEntries(
          { name: 'mark_child_frame', includeChildFrames: true })[0];

        // Child PerformanceMark source should be the child window.
        assert_equals(childEntry.source, childFrame.contentWindow);

        resolve();
      });
    });
    document.body.appendChild(childFrame);
    await loadChildFramePromise;

    // Remove child frame;
    childFrame.parentNode.removeChild(childFrame);

    // Child PerformanceMark source should be null after the child frame is
    // removed.
    assert_equals(childEntry.source, null);
  }, 'PerformanceEntry source is null when the window it points to is removed.')
</script>
